tests: create testutils
